UI tweak for Windows
On Windows we use Gtk.ApplicationWindow instead of Adw.ApplicationWindow which causes an issue where the dialogs that are shown on top don't seem to disable (unfocus) the window/widget below - even if it does. So we manually override the realize and close signals of Gtk.Widget for Adw.Dialog to set the parent widget sensitivity to True or False.